Bug 1132771 - Add a test for reading all moz.build files in filesystem traversal mode;
moz.build files should execute in filesystem traversal mode. Add a test that verifies this is true.

This test performs a brute force filesystem scan to find relevant moz.build files. This can be a little slow. That's unfortunate. But it's a price we need to pay in order to ensure metadata extraction mode continues to work.