Bug 1132771 - API to return moz.build files relevant for a set of paths;
We have an eventual goal to store file-level metadata in moz.build files and to have this metadata "cascade" down directory hierarchies. e.g.  metadata in the root directory will apply to all children directories.

A prerequisite for this feature is a way to query which moz.build files are relevant to a given file. In this patch, we implement an API that returns this information.