Bug 1144249 - fix fetch no-cors mode. r=bkelly
A no-cors mode request should not go through a nsCORSProxyListener,
which can only handle same-origin and cors mode.

The test for no-cors was passing allowOrigin, which was converting it
into a CORS request, which is why it was passing.

I also realized that it is port 8888 and not 8000, so fixed that in some
of the redirect tests.