Bug 1143511 - Correct the usages of .catch() in DOM cache tests; r=bkelly
These tests used .catch() incorrectly, in a way that did not actually
guarantee that the promise is rejected, so in 3 cases the promise
resolved to known failure values (undefined or an empty array) and
this was not being caught as a test failure.

This patch fixes those three places, and also rewrites the .catch()
usage in another case to guarantee that the promise is indeed rejected.