Bug 847287 patch 11 - Only check that we can throttle animations once per refresh cycle (or invalidation of style rule), to make it cheaper to call EnsureStyleRuleFor more than once per refresh cycle. r=birtles
This saves some extra work that we don't need to do.

Mechanically, the patch moves a chunk of code that is around the last
part of the function and converts it to an early return that's slightly
earlier than that last part, thus also including the skipping of the
throttling checks in what we skip for the early return.

I want to do this here since patch 9 introduces a new call to
EnsureStyleRuleFor.