Bug 1112304: Update about:tabcrashed to match the new UX spec. r=dao
Uses in-content styles for the tab crash page and adds an overlay to the favicon
for crashed tabs. Adds support for closing the crashed tab. The strings here
also refer to being able to restore all tabs, that will be implemented and
landed at the same time in bug 1109650 to avoid l10n churn.

Bug 1087618 - Part 2: Add about:welcomeback, sessionrestore, tabcrashed in-content icons. r=jaws