Bug 1159378 - Part 3: Refactor the logic of test_periodic_update.html into a helper script; r=nsm
This helper script will allow us to run the same test across both HTTP
and HTTPS origins.  This patch also uses postMessage instead of a direct
function call on the parent window, because in the HTTPS case, that
window will be cross-origin.  It also moves the code to trigger the
update and get the corresponding updatefound event to
wait_for_update.html, since getting the service worker registration
cross-origin is not possible.