Bug 1157569 part 5 - Convert GetColorAndStyle, GetPaintStyleInfo, and CompareBorders in nsTableFrame to accept writing mode and logical side. r=roc
This changes behavior of border collapse on RTL tables, but it seems
that there is no test coverage on those cases, and those cases should
have been broken before, due to the mix of m{Start,End}side and
physical sides in the methods. Hence this change should be fine, as
long as we are going to fix it in later patches.